+++ packages/icons/src/lib/components/circleHalf.vue
@@ -1,14 +1,5 @@
-<template>
-  <svg v-bind="iconSvgAttrs" :width="size" :height="size">
-    <path
-      fill="currentColor"
-      d="M448 256c0-106-86-192-192-192l0 384c106 0 192-86 192-192zM0 256a256 256 0 1 1 512 0 256 256 0 1 1 -512 0z"
-    />
-  </svg>
-</template>
+<script lang="ts">
+import { createOutlineIcon } from '../internal/outlineIcon';
 
-<script setup lang="ts">
-import { iconSizeDefaults, iconSvgAttrs, type IconSizeProps } from './_shared';
-
-withDefaults(defineProps<IconSizeProps>(), iconSizeDefaults);
+export default createOutlineIcon('circleHalf');
 </script>